fix(dialog): make Dialog modal=false so nested portaled pickers stay clickable
Radix's modal Dialog calls hideOthers(content) and wraps its Overlay in
RemoveScroll, both scoped to the Dialog's own portaled content subtree.
A Select/Combobox popover (e.g. SelectFeature) rendered inside a Dialog
is portaled to document.body as a *separate* subtree, so it gets swept
up by hideOthers' "block pointer-events on everything outside" pass -
the option list renders, but clicks on it don't register (cursor falls
through to whatever's underneath, e.g. showing a text-select cursor
instead of a pointer). This is the same bug class already fixed for
Sheet in #1220, just never applied to Dialog: "Add usage charge" ->
Feature dropdown inside AddSubscriptionChargeDialog was unselectable
for exactly this reason.

Apply the same fix Sheet already uses: modal={false} on the Dialog
root, a hand-rendered pointer-events-none overlay (Radix's own Overlay
renders nothing in non-modal mode), and the outside-dismiss guards so
clicking a portaled dropdown doesn't read as an "outside click" and
close the Dialog out from under it.

Moved useSheetOutsideDismissGuards (and its private helpers) from
Sheet.tsx into modal-scroll-lock.ts so both Sheet and Dialog can share
it; Sheet re-exports it for the one existing external import
(AddEntitlementDrawer).
